Fix rtc calculation

+++ Core/Inc/rtc.h
@@ -6,15 +6,15 @@ typedef struct
     uint8_t hours;
     uint8_t minutes;
     uint8_t seconds;
-    uint16_t subseconds;
+    uint32_t subseconds;
 } RTC_Time_t;
 
 typedef struct
 {
-    uint8_t year;
-    uint8_t month;
-    uint8_t day;
-    uint8_t weekday;
+    uint32_t year;
+    uint32_t month;
+    uint32_t day;
+    uint32_t weekday;
 } RTC_Date_t;
 
 void RTC_Init(void);
